Fix: 25.07.10 가중치 데이터 db 연결 오류 해결

+++ backend/app/schemas/job.py
@@ -10,7 +10,7 @@ class TeamMemberDto(BaseModel):
 
 class WeightDto(BaseModel):
     item: str
-    score: int
+    score: float
 
 
 class ScheduleDto(BaseModel):
